+++ CHANGELOG.md
@@ -3,8 +3,8 @@
 
 1. [](#new)
     * Updated `permissions` field to use new `$grav['permissions']`
-    * DEPRECATED `onAdminRegisterPermissions` event, use `RegisterPermissionsEvent::class` event instead
-    * DEPRECATED `Admin::setPermissions()` and `Admin::addPermissions()`, use `RegisterPermissionsEvent::class` event instead
+    * DEPRECATED `onAdminRegisterPermissions` event, use `PermissionsRegisterEvent::class` event instead
+    * DEPRECATED `Admin::setPermissions()` and `Admin::addPermissions()`, use `PermissionsRegisterEvent::class` event instead
     * DEPRECATED `Admin::getPermissions()`, use `$grav['permissions']->getInstances()` instead
 * [](#improved)
     * Added `field.show_label` and `field.label` display logic from frontend forms